Add new API test script and update .gitignore
- Introduced `api-test.py` for testing text and audio dialogue interactions with the API, including functions for handling responses and saving audio files.
- Updated `.gitignore` to exclude `recordings/`, `third_party/`, and `markitdown/` directories to prevent unnecessary files from being tracked.
- Removed obsolete files: `gui_main.py`, `read_excel.py`, `run_tests.py`, `start_chat.bat`, and a markdown file related to patient Q&A.
